Create README.md
Find a color logo image online and download it to your MATLAB folder. (It can be a logo of a sports team, a
product, a restaurant, or even a country’s flag.) It will be easiest of the logo has very distinct colors (not a
gradual gradient of colors). It will works best if the file is a jpg file. You may want to include that in your
search for a file.
Create a script that will change each color in the logo to a new color. (For example, the red turns blue, the
blue turns yellow, and the yellow turns red.)
The script should then display both the original logo and the modified one.
(It may be fun to take one sports team’s logo and change it to the colors of its rival. For example, change a
Dodgers’ logo so that it has Giants’ colors... Just an idea.)
Submit your script AND the image file in a compressed (zipped) folder.